Refactoring Variable Names

+++ app/src/main/java/org/otdshco/gauges/Params.java
@@ -2,7 +2,7 @@ package org.otdshco.gauges;
 
 public class Params
 {
-    public float roll = 0;
-    public float pitch = 0;
-    public float flightPath = 0;
+    public double roll = 0;
+    public double pitch = 0;
+    public double flightPath = 0;
 }
